refactor: remove messages package in favor of dedicated types
- delete `internal/tui/messages` package and relocate functionality
- move `Animate()` and `FrameMsg` to transitions package
- move `UpdateSlidesMsg` struct from `tui.go` to `slide.go`
- update all references to use the new locations
