+++ internal/tui/transitions/slide_down.go
@@ -7,8 +7,6 @@ import (
 
 	tea "github.com/charmbracelet/bubbletea"
 	"github.com/charmbracelet/harmonica"
-
-	"github.com/ploMP4/kyma/internal/tui/messages"
 )
 
 type slideDown struct {
@@ -54,7 +52,7 @@ func (t slideDown) Update() (Transition, tea.Cmd) {
 		return t, nil
 	}
 
-	return t, messages.Animate(time.Duration(t.fps))
+	return t, Animate(time.Duration(t.fps))
 }
 
 func (t slideDown) View(prev, next string) string {
